+++ src/maximo/provider.js
@@ -1,0 +1,13 @@
+
+
+export default class ServerSourceProvider {
+    constructor(sourceMap) {
+        this.sourceMap = sourceMap;
+
+    }
+    provideTextDocumentContent(uri, token) {
+        let source = this.sourceMap[uri.path];
+        this.sourceMap.delete(uri.path);
+        return source;
+    }
+}